requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.3
Pillow==9.1.1
requests==2.27.1
torch
torchaudio
torchvision
typing-extensions==4.2.0
urllib3==1.26.9
